feat(test-run): it doesnt download production data if it has already been downloaded

+++ tests/run.sh
@@ -10,9 +10,10 @@ if [ -n "$MONGODUMP_BUCKET" ];
 then
     dumps=(`aws s3 --profile $AWS_PROFILE --region $AWS_REGION ls $MONGODUMP_BUCKET/ | awk '{print $4}'`)
     last_dump=${dumps[-1]}
-    aws s3 --profile $AWS_PROFILE --region $AWS_REGION cp $MONGODUMP_BUCKET/${last_dump} /tmp
+    if [ ! -f /tmp/${last_dump} ]; then
+        aws s3 --profile $AWS_PROFILE --region $AWS_REGION cp $MONGODUMP_BUCKET/${last_dump} /tmp
+    fi
     tar -xf /tmp/${last_dump} --directory /tmp
-
     rm -frd /tmp/latest_mongo_dump || true
     mkdir /tmp/latest_mongo_dump
     cp -R /tmp/${last_dump%%.*} /tmp/latest_mongo_dump/dump
